moved channel pooling into own file
